+++ packages/cloud_firestore/cloud_firestore_platform_interface/test/method_channel_field_value_factory_test.dart
@@ -18,7 +18,7 @@ void main() {
   initializeMethodChannel();
   group("$MethodChannelFieldValueFactory()", () {
     setUpAll(() async {
-      await FirebaseCore.instance.initializeApp(
+      await Firebase.initializeApp(
         name: 'testApp',
         options: const FirebaseOptions(
           appId: '1:1234567890:ios:42424242424242',