+++ packages/firebase_storage/test/firebase_storage_test.dart
@@ -21,7 +21,7 @@ void main() {
 
   group('FirebaseStorage', () {
     setUpAll(() async {
-      app = await FirebaseCore.instance.initializeApp(
+      app = await Firebase.initializeApp(
         name: 'testApp',
         options: const FirebaseOptions(
           appId: '1:1234567890:ios:42424242424242',